updated the test cases
