+++ Include/parse_Airports.h
@@ -7,6 +7,6 @@
 //
 
 #include "datafile.h"
-#include "helper_functions.h"
+
 
 int parse_airports(struct AIRPORT** my_airports, char *airport_file);
