+++ src/parse_Airports.c
@@ -7,7 +7,7 @@
 //
 
 #include "../include/datafile.h"
-#include "../include/helper_functions.h"
+
 
 int parse_airports(struct AIRPORT** my_airports, char *airport_file)
 {
