--- test/include/set_airports.h
@@ -1,14 +1,0 @@
-//
-//  main.c
-//  Set_airports
-//
-//  Created by Olumide Akinola on 2020-03-21.
-//  Copyright © 2020 Olumide Akinola. All rights reserved.
-//
-
-
-#include "../../include/data_types.h"
-
-struct AIRPORT** set_airports(int* size);
-
-int parse_airports();
